# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Warden API",
    "version": "0.1.0",
    "description": "Self-hosted game server panel. Bearer JWT from /api/auth/* goes in `Authorization: Bearer <token>`. The live console is a WebSocket (see /ws/servers/{id}/console)."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
...
    "/api/servers": {
      "get": {
        "tags": [
          "servers"
        ],
        "responses": {
          "200": {
            "description": "Server list"
          }
        }
      },
      "post": {
        "tags": [
          "servers"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "game",
                  "name"
                ],
                "properties": {
                  "game": {
                    "type": "string",
                    "example": "minecraft-java"
                  },
                  "name": {
                    "type": "string"
                  },
                  "env": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Unknown game or missing fields"
          }
        }
      }
    },
    "/api/servers/{id}": {
      "get": {
        "tags": [
          "servers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One server"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "tags": [
          "servers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted; container stopped/removed, data dirs removed"
          },
          "403": {
            "description": "Not the owner (non-admin)"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/servers/{id}/{action}": {
      "post": {
        "tags": [
          "servers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "action",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "start",
                "stop",
                "restart",
                "kill"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated server with status"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Driver error"
          }
        }
      }
    },
    "/api/servers/{id}/stats": {
      "get": {
        "tags": [
          "servers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CPU/memory usage. Live container stats when Docker is attached, simulated values (mock:true) otherwise; zeros when stopped."
          },
          "403": {
            "description": "Not the owner (non-admin)"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}